+++ eva/src/index.js
@@ -11,6 +11,7 @@ import VueDraggableResizable from 'vue-draggable-resizable';
 // TODO: что это за модуль, почему ide считает что он не установлен
 import VueCtkDateTimePicker from 'vue-ctk-date-time-picker';
 // подключаем файл с настройками хранилища Vuex
+import moment from 'moment';
 import store from './store/index';
 // подключаем файл с настройками хранилища Vuex (формы)
 // import storeForm from './store/storeForm/index';
@@ -149,8 +150,7 @@ Vue.directive('blur', {
   },
 });
 
-// store.form = storeForm;
-// store.auth = storeAuth;
+moment.locale('ru');
 
 Vue.config.productionTip = true;
 // создаем экземпляр vue где подключаем самое главное переадрасицию,